+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The *alited* provides the rich and comprehensive set of features, some are uniqu
 # LAST CHANGES
 
 
+Version `v1.8.12.2 (10 Dec'25)`
+
+  - CHANGE: Search/Go to Line: non-modal (to switch texts while open) + saved geometry
+  - CHANGE: packages: apave/e_menu 4.8.2
+
+
 Version `v1.8.12.1 (4 Dec'25)`
 
   - BUGFIX: file results in Tools/Check Tcl shown before calculation